+++ packages/timeline-state-resolver/src/integrations/vmix/__tests__/vMixXmlStateParser.spec.ts
@@ -1,5 +1,5 @@
 import { VMixTransitionType } from 'timeline-state-resolver-types'
-import { VMixState } from '../vMixStateDiffer'
+import { VMixAudioBusesState, VMixState } from '../vMixStateDiffer'
 import { VMixXmlStateParser } from '../vMixXmlStateParser'
 import { makeMockVMixXmlState } from './vmixMock'
 import { prefixAddedInput } from './mockState'
@@ -93,15 +93,55 @@ describe('VMixXmlStateParser', () => {
 			playlist: false,
 			multiCorder: false,
 			fullscreen: false,
-			audio: [
-				{
-					volume: 100,
+			audioBuses: {
+				M: {
 					muted: false,
-					meterF1: 0.04211706,
-					meterF2: 0.04211706,
+					volume: 100,
 					headphonesVolume: 74.80521,
 				},
-			],
+				A: {
+					muted: false,
+					volume: 100,
+					sendToMaster: false,
+					solo: false,
+				},
+				B: {
+					muted: false,
+					volume: 78.07491,
+					sendToMaster: false,
+					solo: false,
+				},
+				C: {
+					muted: false,
+					volume: 100,
+					sendToMaster: false,
+					solo: false,
+				},
+				D: {
+					muted: false,
+					volume: 100,
+					sendToMaster: false,
+					solo: false,
+				},
+				E: {
+					muted: true,
+					volume: 100,
+					sendToMaster: false,
+					solo: false,
+				},
+				F: {
+					muted: false,
+					volume: 100,
+					sendToMaster: false,
+					solo: false,
+				},
+				G: {
+					muted: false,
+					volume: 100,
+					sendToMaster: false,
+					solo: false,
+				},
+			} as VMixAudioBusesState, // we're parsing a little more, might be useful down the road
 		})
 	})
 
@@ -264,6 +304,33 @@ describe('VMixXmlStateParser', () => {
 					text: {
 						'TextBlock1.Text': 'SomeText',
 						'AnotherBlock.Text': 'Foo',
+					},
+				},
+			},
+		})
+	})
+
+	it('parses images (titles)', () => {
+		const parser = new VMixXmlStateParser()
+
+		const parsedState = parser.parseVMixState(
+			makeMockVMixXmlState([
+				'<input key="a97b8de1-807a-4c14-8eb9-3de0129b41e3" number="1" type="Capture" title="Cam 0" state="Running" position="0" duration="0" loop="False" muted="False" volume="100" balance="0" solo="False" audiobusses="M" meterF1="0.03034842" meterF2="0.03034842"></input>',
+				`<input key="ca9bc59f-f698-41fe-b17d-1e1743cfee88" number="2" type="GT" title="gfx.gtzip" shortTitle="gfx.gtzip" state="Paused" position="0" duration="0" loop="False" >
+	gfx.gtzip
+	<image index="0" name="SomeImage.Source">image1.png</image>
+	<image index="1" name="AnotherImage.Source">image2.jpg</image>
+</input>`,
+				'<input key="1d70bc59-6517-4571-a0c5-932e30311f01" number="3" type="Capture" title="Cam 2" state="Running" position="0" duration="0" loop="False" muted="False" volume="100" balance="0" solo="False" audiobusses="M" meterF1="0.03034842" meterF2="0.03034842"></input>',
+			])
+		)
+
+		expect(parsedState).toMatchObject<Partial<VMixState>>({
+			existingInputs: {
+				'2': {
+					images: {
+						'SomeImage.Source': 'image1.png',
+						'AnotherImage.Source': 'image2.jpg',
 					},
 				},
 			},