+++ packages/timeline-state-resolver/src/integrations/vmix/vMixCommands.ts
@@ -1,4 +1,4 @@
-import { VMixCommand } from 'timeline-state-resolver-types'
+import { MappingVmixAudioBus, VMixCommand } from 'timeline-state-resolver-types'
 
 export interface VMixStateCommandBase {
 	command: VMixCommand
@@ -51,6 +51,19 @@ export interface VMixStateCommandAudioBusOff extends VMixStateCommandBase {
 	command: VMixCommand.AUDIO_BUS_OFF
 	input: number | string
 	value: string
+}
+export interface VMixStateCommandBusAudioOn extends VMixStateCommandBase {
+	command: VMixCommand.BUS_AUDIO_ON
+	bus: MappingVmixAudioBus['index']
+}
+export interface VMixStateCommandBusAudioOff extends VMixStateCommandBase {
+	command: VMixCommand.BUS_AUDIO_OFF
+	bus: MappingVmixAudioBus['index']
+}
+export interface VMixStateCommandBusVolume extends VMixStateCommandBase {
+	command: VMixCommand.BUS_VOLUME
+	bus: MappingVmixAudioBus['index']
+	value: number
 }
 export interface VMixStateCommandFader extends VMixStateCommandBase {
 	command: VMixCommand.FADER
@@ -219,6 +232,12 @@ export interface VMixStateCommanSelectIndex extends VMixStateCommandBase {
 	input: string | number
 	value: number
 }
+export interface VMixStateCommandSetImage extends VMixStateCommandBase {
+	command: VMixCommand.SET_IMAGE
+	input: string | number
+	fieldName: string
+	value: string
+}
 export type VMixStateCommand =
 	| VMixStateCommandPreviewInput
 	| VMixStateCommandTransition
@@ -230,6 +249,9 @@ export type VMixStateCommand =
 	| VMixStateCommandAudioAutoOff
 	| VMixStateCommandAudioBusOn
 	| VMixStateCommandAudioBusOff
+	| VMixStateCommandBusAudioOn
+	| VMixStateCommandBusAudioOff
+	| VMixStateCommandBusVolume
 	| VMixStateCommandFader
 	| VMixStateCommandSetZoom
 	| VMixStateCommandSetPanX
@@ -267,6 +289,7 @@ export type VMixStateCommand =
 	| VMixStateCommandSetText
 	| VMixStateCommandBrowserNavigate
 	| VMixStateCommanSelectIndex
+	| VMixStateCommandSetImage
 
 export enum CommandContext {
 	None = 'none',