+++ packages/timeline-state-resolver/src/integrations/vmix/vMixStateDiffer.ts
@@ -7,6 +7,8 @@ import {
 	VMixTransitionType,
 	VMixLayer,
 	VMixText,
+	VMixImages,
+	MappingVmixAudioBus,
 } from 'timeline-state-resolver-types'
 import { CommandContext, VMixStateCommandWithContext } from './vMixCommands'
 import _ = require('underscore')
@@ -43,7 +45,7 @@ export interface VMixState {
 	playlist: boolean
 	multiCorder: boolean
 	fullscreen: boolean
-	audio: VMixAudioChannel[]
+	audioBuses: VMixAudioBusesState
 }
 
 interface VMixOutputsState {
@@ -56,6 +58,10 @@ interface VMixOutputsState {
 	Fullscreen: VMixOutput | undefined
 	Fullscreen2: VMixOutput | undefined
 }
+
+export type VMixAudioBusesState = {
+	M: VMixAudioBusBase | undefined
+} & Record<'A' | 'B' | 'C' | 'D' | 'E' | 'F' | 'G', VMixAudioBusBase | undefined>
 
 export interface VMixMix {
 	number: number
@@ -82,6 +88,7 @@ export interface VMixInput {
 	text?: VMixText
 	url?: string
 	index?: number
+	images?: VMixImages
 }
 
 export interface VMixInputAudio {
@@ -105,11 +112,17 @@ export interface VMixOverlay {
 	input: string | number | undefined
 }
 
-export interface VMixAudioChannel {
+export interface VMixAudioBusBase {
 	volume: number
 	muted: boolean
-	meterF1: number
-	meterF2: number
+}
+
+export interface VMixAudioRegularBus extends VMixAudioBusBase {
+	solo: boolean
+	sendToMaster: boolean
+}
+
+export interface VMixAudioMasterBus extends VMixAudioBusBase {
 	headphonesVolume: number
 }
 
@@ -118,7 +131,14 @@ interface PreAndPostTransitionCommands {
 	postTransitionCommands: Array<VMixStateCommandWithContext>
 }
 
-export class VMixStateDiffer {
+export interface VMixDefaultStateFactory {
+	getDefaultState: () => VMixStateExtended
+	getDefaultInputState: (inputIndex: number | string | undefined) => VMixInput
+	getDefaultInputAudioState: (inputIndex: number | string | undefined) => VMixInputAudio
+	getDefaultAudioBusState: () => VMixAudioBusBase
+}
+
+export class VMixStateDiffer implements VMixDefaultStateFactory {
 	private inputHandler: VMixInputHandler
 
 	constructor(
@@ -140,6 +160,7 @@ export class VMixStateDiffer {
 		commands = commands.concat(this._resolveOverlaysState(oldVMixState, newVMixState))
 		commands = commands.concat(inputCommands.postTransitionCommands)
 		commands = commands.concat(this._resolveInputsAudioState(oldVMixState, newVMixState))
+		commands = commands.concat(this._resolveAudioBusesState(oldVMixState.reportedState, newVMixState.reportedState))
 		commands = commands.concat(this._resolveRecordingState(oldVMixState.reportedState, newVMixState.reportedState))
 		commands = commands.concat(this._resolveStreamingState(oldVMixState.reportedState, newVMixState.reportedState))
 		commands = commands.concat(this._resolveExternalState(oldVMixState.reportedState, newVMixState.reportedState))
@@ -171,7 +192,16 @@ export class VMixStateDiffer {
 				playlist: false,
 				multiCorder: false,
 				fullscreen: false,
-				audio: [],
+				audioBuses: {
+					M: undefined,
+					A: undefined,
+					B: undefined,
+					C: undefined,
+					D: undefined,
+					E: undefined,
+					F: undefined,
+					G: undefined,
+				},
 			},
 			outputs: {
 				'2': undefined,
@@ -211,6 +241,13 @@ export class VMixStateDiffer {
 			fade: 0,
 			audioBuses: 'M',
 			audioAuto: true,
+		}
+	}
+
+	getDefaultAudioBusState(): VMixAudioBusBase {
+		return {
+			muted: true,
+			volume: 100,
 		}
 	}
 
@@ -625,6 +662,22 @@ export class VMixStateDiffer {
 				timelineId: '',
 			})
 		}
+		if (input.images !== undefined) {
+			for (const [fieldName, value] of Object.entries<string>(input.images)) {
+				if (oldInput?.images?.[fieldName] !== value) {
+					commands.push({
+						command: {
+							command: VMixCommand.SET_IMAGE,
+							input: key,
+							value,
+							fieldName,
+						},
+						context: CommandContext.None,
+						timelineId: '',
+					})
+				}
+			}
+		}
 		return { preTransitionCommands, postTransitionCommands }
 	}
 
@@ -779,6 +832,52 @@ export class VMixStateDiffer {
 				this.inputHandler.removeInput(time, input)
 			}
 		)
+		return commands
+	}
+
+	private _resolveAudioBusesState(
+		oldVMixState: VMixState,
+		newVMixState: VMixState
+	): Array<VMixStateCommandWithContext> {
+		const commands: Array<VMixStateCommandWithContext> = []
+		for (const [index, bus] of Object.entries<VMixAudioBusBase | undefined>(newVMixState.audioBuses)) {
+			const busName = index as MappingVmixAudioBus['index']
+			if (!bus) continue
+			const oldBus = oldVMixState.audioBuses[index as keyof VMixAudioBusesState]
+			// probably makes sense to do this before updating volume:
+			if (bus.muted && oldBus?.muted !== bus.muted) {
+				commands.push({
+					command: {
+						command: VMixCommand.BUS_AUDIO_OFF,
+						bus: busName,
+					},
+					context: CommandContext.None,
+					timelineId: '',
+				})
+			}
+			if (oldBus?.volume !== bus.volume) {
+				commands.push({
+					command: {
+						command: VMixCommand.BUS_VOLUME,
+						bus: busName,
+						value: bus.volume,
+					},
+					context: CommandContext.None,
+					timelineId: '',
+				})
+			}
+			// probably makes sense to do this after updating volume:
+			if (!bus.muted && oldBus?.muted !== bus.muted) {
+				commands.push({
+					command: {
+						command: VMixCommand.BUS_AUDIO_ON,
+						bus: busName,
+					},
+					context: CommandContext.None,
+					timelineId: '',
+				})
+			}
+		}
 		return commands
 	}